move typeGuards to utils
