+++ packages/core/src/events.ts
@@ -1,28 +1,17 @@
 import {
+  isJsonRpcRequest,
+  isJsonRpcResponseSuccess,
+  isJsonRpcResponseError,
+  isInternalEvent,
+  isReservedEvent
+} from '@walletconnect/utils'
+import {
   IInternalEvent,
   IJsonRpcResponseSuccess,
   IJsonRpcResponseError,
   IJsonRpcRequest,
   IEventEmitter
 } from '@walletconnect/types'
-
-// -- typeChecks ----------------------------------------------------------- //
-
-function isJsonRpcRequest (object: any): object is IJsonRpcRequest {
-  return 'method' in object
-}
-
-function isJsonRpcResponseSuccess (object: any): object is IJsonRpcResponseSuccess {
-  return 'result' in object
-}
-
-function isJsonRpcResponseError (object: any): object is IJsonRpcResponseError {
-  return 'error' in object
-}
-
-function isInternalEvent (object: any): object is IInternalEvent {
-  return 'event' in object
-}
 
 // -- EventManager --------------------------------------------------------- //
 
@@ -49,7 +38,10 @@ class EventManager {
 
     if (isJsonRpcRequest(payload)) {
       event = payload.method
-    } else if (isJsonRpcResponseSuccess(payload) || isJsonRpcResponseError(payload)) {
+    } else if (
+      isJsonRpcResponseSuccess(payload) ||
+      isJsonRpcResponseError(payload)
+    ) {
       event = `response:${payload.id}`
     } else if (isInternalEvent(payload)) {
       event = payload.event
@@ -63,16 +55,6 @@ class EventManager {
       )
     }
 
-    function isReservedEvent (event: string) {
-      const reservedEvents = [
-        'session_request',
-        'session_update',
-        'exchange_key',
-        'connect',
-        'disconnect'
-      ]
-      return reservedEvents.includes(event) || event.startsWith('wc_')
-    }
     if ((!eventEmitters || !eventEmitters.length) && !isReservedEvent(event)) {
       eventEmitters = this._eventEmitters.filter(
         (eventEmitter: IEventEmitter) => eventEmitter.event === 'call_request'
