+++ packages/utils/README.md
@@ -54,4 +54,12 @@ function parsePersonalSign (params: string[]): string[]
 function parseTransactionData (txData: Partial<ITxData>): Partial<ITxData>
 
 function formatRpcError (error: Partial<IJsonRpcErrorMessage>): { code: number; message: string }
+
+function isJsonRpcSubscription (object: any): boolean
+function isJsonRpcRequest (object: any): boolean
+function isJsonRpcResponseSuccess (object: any): boolean
+function isJsonRpcResponseError (object: any): boolean
+function isInternalEvent (object: any): boolean
+function isWalletConnectSession (object: any): boolean
+function isReservedEvent (event: string): boolean
 ```
